requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
idna==3.4
psycopg2-binary==2.9.7
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
stripe==6.5.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5